requirements.txt
streamlit>=1.33.0
opencv-python-headless>=4.9.0.80
numpy>=1.26.4
Pillow>=10.3.0
python-dotenv>=1.0.1
google-generativeai>=0.3.2
openai>=1.30.1
easyocr>=1.7.1
langdetect>=1.0.9
language-tool-python>=2.7.1
pdf2image>=1.17.0
poppler-utils
